删除不必要的代码

+++ pub_app.py
@@ -23,12 +23,11 @@ app.config.from_pyfile(CONFIG_FILE)
 # 自动关闭数据库连接
 @app.teardown_appcontext
 def close_db(exception=None):
-    if exception is not None:
-        print('++++' + str(exception) + '++++')
-    db.remove()
+    if db is not None:
+        db.remove()
 
 @app.teardown_request
-def teardown_request(exception):
+def teardown_request(exception=None):
     if db is not None:
         db.close()
